refactor hazecast port and spring cloud utils for network
