+++ tesla-common/src/main/java/io/github/tesla/common/utils/InetUtilsProperties.java
@@ -1,16 +1,4 @@
-/*
- * Copyright 2014-2017 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-package io.github.tesla.gateway.config.eureka.util;
+package io.github.tesla.common.utils;
 
 import java.net.InetAddress;
 import java.util.ArrayList;
@@ -19,17 +7,10 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
-/**
- * @author liushiming
- * @version InetUtilsProperties.java, v 0.0.1 2018年5月4日 下午6:47:29 liushiming
- */
 @ConfigurationProperties(InetUtilsProperties.PREFIX)
 public class InetUtilsProperties {
-    public static final String PREFIX = "spring.cloud.inetutils";
 
-    public static String getPREFIX() {
-        return PREFIX;
-    }
+    public static final String PREFIX = "spring.cloud.inetutils";
 
     /**
      * The default hostname. Used in case of errors.
@@ -62,51 +43,56 @@ public class InetUtilsProperties {
      */
     private List<String> preferredNetworks = new ArrayList<>();
 
+    public static String getPREFIX() {
+        return PREFIX;
+    }
+
     public String getDefaultHostname() {
         return defaultHostname;
-    }
-
-    public String getDefaultIpAddress() {
-        return defaultIpAddress;
-    }
-
-    public List<String> getIgnoredInterfaces() {
-        return ignoredInterfaces;
-    }
-
-    public List<String> getPreferredNetworks() {
-        return preferredNetworks;
-    }
-
-    public int getTimeoutSeconds() {
-        return timeoutSeconds;
-    }
-
-    public boolean isUseOnlySiteLocalInterfaces() {
-        return useOnlySiteLocalInterfaces;
     }
 
     public void setDefaultHostname(String defaultHostname) {
         this.defaultHostname = defaultHostname;
     }
 
+    public String getDefaultIpAddress() {
+        return defaultIpAddress;
+    }
+
     public void setDefaultIpAddress(String defaultIpAddress) {
         this.defaultIpAddress = defaultIpAddress;
     }
 
-    public void setIgnoredInterfaces(List<String> ignoredInterfaces) {
-        this.ignoredInterfaces = ignoredInterfaces;
-    }
-
-    public void setPreferredNetworks(List<String> preferredNetworks) {
-        this.preferredNetworks = preferredNetworks;
+    public int getTimeoutSeconds() {
+        return timeoutSeconds;
     }
 
     public void setTimeoutSeconds(int timeoutSeconds) {
         this.timeoutSeconds = timeoutSeconds;
     }
 
+    public List<String> getIgnoredInterfaces() {
+        return ignoredInterfaces;
+    }
+
+    public void setIgnoredInterfaces(List<String> ignoredInterfaces) {
+        this.ignoredInterfaces = ignoredInterfaces;
+    }
+
+    public boolean isUseOnlySiteLocalInterfaces() {
+        return useOnlySiteLocalInterfaces;
+    }
+
     public void setUseOnlySiteLocalInterfaces(boolean useOnlySiteLocalInterfaces) {
         this.useOnlySiteLocalInterfaces = useOnlySiteLocalInterfaces;
     }
+
+    public List<String> getPreferredNetworks() {
+        return preferredNetworks;
+    }
+
+    public void setPreferredNetworks(List<String> preferredNetworks) {
+        this.preferredNetworks = preferredNetworks;
+    }
+
 }
