+++ tesla-filter/src/main/java/io/github/tesla/filter/service/plugin/request/AuthRequestPlugin.java
@@ -18,11 +18,6 @@ import io.netty.handler.codec.http.HttpObject;
 import io.netty.handler.codec.http.HttpResponse;
 import io.netty.handler.codec.http.HttpResponseStatus;
 
-/**
- * @author: zhangzhiping
- * @date: 2018/11/20 14:32
- * @description: auth filter入口，下分三种校验
- */
 @ServiceRequestPlugin(filterType = "AuthRequestPlugin", definitionClazz = AuthCommonDefinition.class, filterOrder = 3,
     filterName = "权限验证插件")
 public class AuthRequestPlugin extends AbstractRequestPlugin {
@@ -34,7 +29,7 @@ public class AuthRequestPlugin extends AbstractRequestPlugin {
 
     @Override
     public HttpResponse doFilter(NettyHttpServletRequest servletRequest, HttpObject realHttpObject,
-        String  filterParam) {
+        String filterParam) {
         AuthCommonDefinition authDefinition = JsonUtils.json2Definition(filterParam, AuthCommonDefinition.class);
         // 不启用auth，则返回
         if (authDefinition == null || YesOrNoEnum.NO.getCode().equals(authDefinition.getEnabled())) {
@@ -50,6 +45,64 @@ public class AuthRequestPlugin extends AbstractRequestPlugin {
             PluginUtil.writeFilterLog(AuthRequestPlugin.class, e.getMessage(), e);
             return PluginUtil.createResponse(HttpResponseStatus.INTERNAL_SERVER_ERROR, servletRequest.getNettyRequest(),
                 "auth filter error " + e.getMessage());
+        }
+    }
+
+    protected static class ResponseMessage {
+        public enum MesageCodeType {
+
+            OSG001("Token expire"), OSG002("Wrong Token"), OSG003(" Token has not exist");
+
+            private String code;
+
+            MesageCodeType(String code) {
+                this.code = code;
+            }
+
+            public String getCode() {
+                return code;
+            }
+
+            public void setCode(String code) {
+                this.code = code;
+            }
+
+        }
+
+        private MesageCodeType code;
+
+        private String message;
+
+        private String content;
+
+        public ResponseMessage(MesageCodeType code, String message) {
+            super();
+            this.code = code;
+            this.message = message;
+        }
+
+        public MesageCodeType getCode() {
+            return code;
+        }
+
+        public String getContent() {
+            return content;
+        }
+
+        public String getMessage() {
+            return message;
+        }
+
+        public void setCode(MesageCodeType code) {
+            this.code = code;
+        }
+
+        public void setContent(String content) {
+            this.content = content;
+        }
+
+        public void setMessage(String message) {
+            this.message = message;
         }
     }
 }
