+++ tesla-gateway/src/main/java/io/github/tesla/gateway/config/eureka/EurekaInstanceConfigBean.java
@@ -26,9 +26,9 @@ import com.netflix.appinfo.EurekaInstanceConfig;
 import com.netflix.appinfo.InstanceInfo.InstanceStatus;
 import com.netflix.appinfo.MyDataCenterInfo;
 
-import io.github.tesla.filter.support.springcloud.SpringEnvironmentUtil;
-import io.github.tesla.gateway.config.eureka.util.InetUtils;
-import io.github.tesla.gateway.config.eureka.util.InetUtils.HostInfo;
+import io.github.tesla.common.utils.InetUtils;
+import io.github.tesla.common.utils.InetUtils.HostInfo;
+import io.github.tesla.gateway.config.eureka.util.SpringEnvironmentUtil;
 
 /**
  * @author liushiming
@@ -249,8 +249,7 @@ public class EurekaInstanceConfigBean implements EurekaInstanceConfig, Environme
         return getHostName(false);
     }
 
-    private EurekaInstanceConfigBean() {
-    }
+    private EurekaInstanceConfigBean() {}
 
     public EurekaInstanceConfigBean(InetUtils inetUtils) {
