+++ tesla-gateway/src/main/java/io/github/tesla/gateway/config/eureka/util/SpringEnvironmentUtil.java
@@ -1,4 +1,4 @@
-package io.github.tesla.filter.support.springcloud;
+package io.github.tesla.gateway.config.eureka.util;
 
 import org.springframework.core.env.Environment;
